Library

Library

Sampling

Non-Parametric Sampling

Bootstrap.BasicSampling — Type.

Basic Sampling

BasicSampling(1000)

Bootstrap.BalancedSampling — Type.

Balanced Sampling

BalancedSampling(1000)

Bootstrap.ExactSampling — Type.

Exact Sampling

ExactSampling()

Bootstrap.AntitheticSampling — Type.

Antithetic Sampling

AntitheticSampling(1000)

Bootstrap.MaximumEntropySampling — Type.

Maximum Entropy Sampling

MaximumEntropySampling(100)
maximumEntropySampling(100, MaximumEntropyCache())

NOTE: Implementation based off pymeboot as the original R package is GPL licensed.

Parametric Sampling

Bootstrap.ResidualSampling — Type.

Residual Sampling

ResidualSampling(1000)

Bootstrap.WildSampling — Type.

Wild Sampling

WildSampling(1000, rademacher)
WildSampling(1000, mammen)

Accessors

Bootstrap.nrun — Function.

Number of samples drawn from a bootstrap sampling

bs = BasicSampling(1000)
nrun(bs)

# output

1000

Bootstrap.statistic — Function.

Return the statistic function of a BootstrapSample

bs = bootstrap(randn(20), mean, BasicSampling(100))
statistic(bs)

Bootstrap.noise — Function.

Return the noise function of a wild bootstrap sampling

Confidence Intervals

Bootstrap.BasicConfInt — Type.

Basic Confidence Interval

BasicConfInt(0.95)

Bootstrap.PercentileConfInt — Type.

Percentile Confidence Interval

PercentileConfInt(0.95)

Bootstrap.NormalConfInt — Type.

Normal Confidence Interval

NormalConfInt(0.95)

Bootstrap.StudentConfInt — Type.

Student's Confidence Interval

StudentConfInt(0.95)

Bootstrap.BCaConfInt — Type.

Bias-Corrected and Accelerated (BCa) Confidence Interval

BCaConfInt(0.95)